backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pydantic==2.5.0
python-magic==0.4.27
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.220
opencv-python-headless==4.8.1.78
slowapi==0.1.9
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI
